Create hook to handle routes
